# Improve speed for pipedgzip reader.