@rhpvorderman 2019-08-12
For context I did a hyperfine run of `pigz -cd` itself:
```
hyperfine -w 3 "pigz -cd ~/test/big.fastq.gz >> /dev/null"
Benchmark #1: pigz -cd ~/test/big.fastq.gz >> /dev/null
  Time (mean ± σ):      4.310 s ±  0.097 s    [User: 6.783 s, System: 0.441 s]
  Range (min … max):    4.144 s …  4.516 s    10 runs
 ```
Using pigz alone it uses 160% cpu or 1.6 core, therefore the real time is less than the cpu time.
pigz takes on average 6.8 cpu seconds to decompress this file.
Before this PR: 15.1 - 6.8  = 8.3 cpu seconds python overhead
After this PR: 12.6 - 6.8 = 5.8 cpu seconds python overhead.
5.8 / 8.3 ~= 0.7. A 30% reduction in python overhead is achieved.

Given that the python CPU overhead equals the real time it shows that python is the limiting factor still (I can also see this on my system monitor, python reaches 100% of one core all the time). `pigz` decompresses faster than python can iterate over lines.