@marcelm 2019-08-12
Thanks, this is interesting!

I had to do some digging why the original code was written this way. After all, your patch removes the `self.process.wait()` and `self._raise_if_error()` lines, which are obviously meant for error handling. This is the original commit in Cutadapt: https://github.com/marcelm/cutadapt/commit/63b787617e1564e4838b3b4692789f0ddfd3b523

That commit introduces the `PipedGzipReader` class. Previously, I just used `subprocess.Popen(...)` directly, but this hid errors when reading truncated gzip files.

Interestingly, all tests pass with your patch, in particular even those that check reading from truncated gzip files raises an error!

Here is one of those tests (abbreviated):
```
f = xopen("truncated.gz", "r")
for line in f:
    pass
f.close()
```
As it turns out, the current version will correctly crash on one of the `for line in f` iterations, and the changed version will not do so. However, your version will crash on `f.close()`, so the test still passes.

So this change will be an inconvenience for those users who do not close their input files, but one can argue that it’s already a bug in their code.

One alternative is to use a `yield from self._file` instead, and then the error checking can be left in. Do you think you could run your time measurements for that as well? If the runtime is the same, I’d prefer this, but otherwise, I’m probably fine with merging it as-is.