@rhpvorderman 2019-08-12
I realized that no errors would be checked in the iter method, but since the checking self._raise_if_error() was also done in the close() method I figured that the checking would be done anyway. If files are not properly closed that leads to other problems as well.

I tested the following snippet:

```python
    def __iter__(self):
        yield from self._file
        self.process.wait()
        self._raise_if_error()
```
Benchmark:
```
hyperfine -w 3 "python3 test.py ~/test/big.fastq.gz"
Benchmark #1: python3 test.py ~/test/big.fastq.gz
  Time (mean ± σ):      7.727 s ±  0.265 s    [User: 14.558 s, System: 1.881 s]
  Range (min … max):    7.257 s …  8.138 s    10 runs
```
So that is considerably slower than returning self._file.