@marcelm 2019-08-12
Thanks for running the extra benchmark! This is a significant improvement, thanks a lot.