@rhpvorderman 2019-08-14
@marcelm, could you make a new release? It is nice to have the extra speed available on pypi. Thanks!