@marcelm 2019-08-14
Sure, I just tagged 0.8.0.

However, I’m a bit worried that something is now broken because [of this test failure](https://travis-ci.org/marcelm/xopen/builds/571788800). This is the test run for the commit where I only updated the changelog, so it should pass. The next build does indeed pass, which is why the automatic release to PyPI went through. Possibly this is a non-deterministic failure.