@marcelm 2019-08-14
I don’t have time at the moment, but have opened #16 to keep track of this.